Reformat, phone number changes, and translations
Made the file adhere to PEP standards regarding spacing.

Phone number is now allowed to be bigger to support foreign phone #s.
This is sort of dirty, because phone numbers should have their country
code associated to perform some sort of validation... But this is way
too low in priority, as this field is merely informative and is rarely
used IRL.

Marked more strings as needing to be translated.