Improve patch behaviour for slots/authors
Previously, patching slots or authors was unspecified. It is now 100%
explicit, and works as expected.

Creation and Update of the serialized object was improved by adding a
transaction.atomic decorator - this way, deletes and saves don't
autocommit, which means that, on error, they are rolled back. This
simplifies greatly the code needed to be written, as it is written for
a situation with no errors, and if there are errors, the DB stays
intact.