Add invitations for users
When a user is registered, there shouldn't be a password set. He/she
should be invited to manually set his/her password on the frontend. For
this, an invite is created when the user is created through the service,
which contains a token and other info required to make sure the user
matches. After some validation, the chosen password is set.

The invite is sent through the email associated to the user, and is sent
as an HTML email or, if the email client does not support it, as plain
text.

The invite is only sent if the user is active: we want to be able to
register old users which may not be participating in the Radio anymore,
and so it doesn't make sense to send them an email to set a password.