Add invite accept endpoint and test
This commit required a slight change in the user manager to ensure that
it is always saved without a password, making the action of accepting an
invite the only way to set a password.

User also has a field has_registered to be able to easily view at a
glance the authors who have already accepted their invites.