Add invite resend and list endpoint and tests
Invite resend is to be used when a user claims to not have received an
email, or can't find it somehow... This avoids us having to recreate an
account for that user so the invite is sent again.

Listing invites is useful to be able to view at a glance the invites
that may need to be resent.