Add password reset, fix tests
The project now supports password resetting, courtesy of a new library.

Tests were broken due to the addition of authentication protection on
the views. They were fixed and now tolerate this by creating a new user,
getting the token associated to this new user, and using it for headers
in every request.

The BASE_URL property was changed to BASE_FRONTEND_URL, because this is
used only for emails, and the user should open the frontend page, which
is in turn responsible for the proper API call.