Add generation of access keys for studio
This commit includes a new app, used only for generating a new pdf file
containing all of our members, as well as some introductory test. This
pdf is meant to be used as a reference sheet for the security services
at Técnico, to lent a key to enter our studio.

The generation works well (provided that everything is properly
installed). WIP is how to serve the PDF to the user with this REST
framework.