Add hashing to pdf generation, refactor service, info view
For security reasons, we want to make sure we keep a trail of all pdfs
generated, even if we discard previously generated PDFs as we generate
new ones. They should always be accessible on the email, but by hashing
the PDF, we can ensure that the file on the email is, indeed, the file
produced by the server.

The KeyDocService was refactored to make the main method shorter, and
delegate tasks to other helper methods.

The new view makes it easy to find out a few stats about the latest
doc generated, for future frontend validation and display.