Add files for Docker
This is a commit regarding Dockerization of the project - in an early
stage.
It runs the manage.py server and is binded to the backend directory, so
any changes in the code instantly affect the container.

In the future, this should run the frontend, as well as celery, ffmpeg,
nginx, gunicorn... Some work to be done.