Change key doc pdf generation for docker, add start script
There were some problems when running key generation inside docker.

* The Ubuntu font can't seem to be installed via apt, so it can't be
used in the .md. We could use the default font on the Eisvogel template,
but:
* The Eisvogel template can only be used if the package tex-live-fonts-
extra is installed, which occupies some whopping 500mb (due to the need
of the font sourcesanspro- which does look good but apparently cannot be
installed alone)

Due to this, defaults for pandoc generation are assumed as we try to
find a way to fix these problems (namely, installing the Ubuntu font).