Add celery and flower to docker services, rollback keys
This commit introduces a reorganization of the docker architecture to
support Celery + Flower for asynchronous tasks.
The base backend image is the same for the 3 services (backend + celery
+ flower), but the last two run a different command. This image was also
changed to no longer being based on the python image: it's now based on
the much more recent Ubuntu20 image, giving us more up to date packages
and allowing the Key Doc generation to work as intended (with the ubuntu
font and eisvogel template).

Further improvements can be made in terms of disk spaced used by this
project, on Docker.