Add emission download from archive
This introduces the ability to download a previously uploaded emission.

The problem with the current approach is that it is too hardcoded - we
are making lots of assumptions regarding the path on the archive server,
filename, etc... It would maybe be better to have a microservice running
on the archive, responsible for this management, where we could just ask
for a file for a given date, and for a given program, and it would
stream it to us.
The advantage of the current approach is that we have less things to
worry about when setting up or fixing bugs: just run Radiologo and
everything works. The trade offs of both approaches must be further
analyzed.