Fix bug when checking archive for file
The stat method throws an IOError when it cannot find the file, and that
is precisely the check we want to make, since we allow the program to
proceed as normal if it happens. If the file can be found, stat returns
an object for analysis, and that is when we must throw an exception.