Add basic stats of uploads
These stats list if a program should have been uploaded for the past
week, and on which days it was or wasn't uploaded.

Program also has a new method for building a filename for the archive.
Other places where the filename is manually built must use this new
method (WIP), making any possible change much easier.