Refactor processing service to receive program pk
This refactor is a 'hack' to pass a program to the processing service.
Since the processing is called through celery, and celery does not like
passing objects on task arguments (due to serialization), we pass the
program's pk, and the service gets the object using the pk. This way,
we avoid having to pass lots of arguments for the program's name and
other fields, and are able to use methods only accessible through the
program instance (such as one to get the output filename for processing)