Add radiologo alignment, fix rss upload task
This commit makes Radiologo responsible for aligning the emissions, by
moving them from our archive server to the upload/emission server.

It also fixes the rss upload task to run the FeedService for every
active program with an active rss feed status, instead of the task
having to receive the pk of the program to run. This makes the task
much more suitable for scheduled (daily or whenever) running.

Closes #4.