Add reruns of programs
A rerun Slot marks a slot where a program repeats a previous emission,
and is simply a flag in a normal Slot object.
There is one rerun slot for each normal slot. A program cannot be
uploaded to its rerun slot.
This makes the schedule more accurate and similar to our current
schedule look, and may allow for future automagic management of reruns.

Some changes were required to ensure users can't upload to emission
slots.