Add permissions
This commit introduces permissions protecting the views. We opted to
write "group permissions", which cover people belonging to a specific
department or having a specific role, and also have a RW (read-write) or
R (read) variant. This way, if people from a certain department start to
be able to have write access to a certain resource, it's just a matter
of going to the view and changing the permission class from R to RW -
the maintenance of permissions becomes much easier.

Closes #11 (although not in the way the issue solves the problem)