Added feed url validation with parsing
When scheduling uploads with a new feed url, it's nice to have some feedback that the URL is valid and reachable, so it won't fail later at an unscheduled time.
Besides just checking that the URL has a valid syntax, the software now attempts to download the feed and to parse it to check if it is valid.
It gives the user some new feedback by returning the name of the feed taken from parsing the contents at the specified URL.