Implement basic login functionality
When logging in, the instance of the logged in user is saved in the
store, for easy access of the user's programs and other info. The JWT
token is also saved, and sent on every Axios request.
The whole store is saved in local storage for persistence across page
refreshes.

It is missing finer permission control over each page, and testing to
ensure that security and redirects in case of access denied or expired
session work.