Add upload of files to frontend
This commit makes possible the upload of files on the frontend. It works
similarly to the previous Radiólogo version, with some slight UI
changes.

The interface for upload could be slightly improved, because currently,
the center card changes sizing according to the content, and it would be
better to stay the same size for the different steps of the upload.
However, it works, and by disabling the calendar when the upload is
finished, we force a refresh of the page if the user wants to upload
something else without us having to have more work setting the page
back to its starting state.