UI improvements in sizing and list, fix bug
The list of the user's programs, both in the drawer and in the top bar,
was not fully consistent with the rest of the style of the website. It
was now changed to use more "native" vuetify components, aka proper
lists instead of thin buttons.

Wording of the long "download" button the ArchiveView was shortened to
better fit mobile screens while keeping the meaning of the button.

The router-view tag in the App.vue file now has a unique key prop. This
was added to ensure that the same page, but with a different URL, is
rendered when the URL changes.