Add table view to archive, download bar
A table view was added to the central card, for those who might prefer
a more concise list of the uploads.

Since files are downloaded and then instantly given to the user, without
the browser natively indicating the progress, a new progress bar was
added for clarity of progress. When the file is being downloaded, the
global infinite loading bar is disabled, to avoid potential confusion
with 2 loading bars.