UI improvements in upload card, top bar wording
This commit offers more in terms of consistency and look/feel of the
upload card. The file picker is now made up of a button instead of a
vuetify file picker. This gives us greater flexibility in terms of
customization, and looks subjectively better.

Top bar had the wording for the user's programs menu changed to clear up
any possible confusion between the user's programs and the
administration of all the programs in the Radio.