Final corrections and migrations for DB stats
This completes the corrections for running DB stats. These will run adding an entry for every hour
and for every program that airs at Rádio Zero.
Authors will soon have the endpoint and interface to see the evolution of listeners during the episode,
including important information of when they left.