Add slots blocking in interface
When selecting a slot for a program, it is important to only allow the
user to select a slot that does not interfere with other slots.
Therefore, the backend calculates the slots that are free, for the given
program/duration/weekdays, and the frontend parses that to build the
selection of slots.